Move subscribe into BlendLoopState.
This helps avoid the '_RestrictContext' warning that comes with attaching things
to bpy.context in the register() function. Blender doesn't want you to be
attaching things during register(), you should attach things when your add-on is
actually invoked. This commit moves with BlendLoopMessageSubscriber
initialization into the loop_start() function that is invoked by an operator.